테스트 코드: REST CONTROLLER 패키지 분기에 따른 테스트 코드 변경
