+++ src/test/java/com/gunyoung/info/controller/ContentControllerTest.java
@@ -1,7 +1,6 @@
 package com.gunyoung.info.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
@@ -321,129 +320,5 @@ public class ContentControllerTest {
 		Long contentId = space.getContents().get(0).getId();
 		mockMvc.perform(get("/space/updatecontent/" + contentId.intValue()))
 				.andExpect(view().name("updateContent"));
-	}
-	
-	/*
-	 *  - 대상 메소드: 
-	 *  	@RequestMapping(value="/space/updatecontent/{id}", method= RequestMethod.POST) 
-	 *		public ModelAndView updateContentPost(@PathVariable long id, @ModelAttribute("formModel") ContentDTO contentDto, ModelAndView mav)
-	 */
-	
-	@WithMockUser(username="test@google.com", roles= {"USER"})
-	@Test
-	@DisplayName("콘텐트 업데이트 POST (실패-입력된 id에 해당하는 content가 DB에 없을때")
-	public void updateContentPostNonExist() throws Exception{
-		MultiValueMap<String, String> map = new LinkedMultiValueMap<>();
-		map.add("title", "changed title"); 
-		map.add("contributors", "changed test contributors");
-		map.add("contents", "changed test contents");
-		
-		mockMvc.perform(post("/space/updatecontent/100")
-				.params(map))
-				.andExpect(status().isNoContent());
-	}
-	
-	@WithMockUser(username="second@naver.com", roles={"USER"})
-	@Test
-	@Transactional
-	@DisplayName("콘텐트 업데이트 POST(실패-해당 id의 콘텐트가 현재 접속자의 것이 아닐때)")
-	public void updateContentPostWrongUser() throws Exception {
-		
-		Person p = personService.findByEmail("test@google.com");
-		Space space = p.getSpace();
-		Content content = space.getContents().get(0);
-		Long contentId = content.getId();
-		String title = content.getTitle();
-		
-		MultiValueMap<String, String> map = new LinkedMultiValueMap<>();
-		map.add("hostEmail", "test@google.com");
-		map.add("title", "changed title"); 
-		map.add("contributors", "changed test contributors");
-		map.add("contents", "changed test contents");
-		
-		mockMvc.perform(post("/space/updatecontent/" + contentId.intValue())
-				.params(map))
-				.andExpect(status().isBadRequest());
-		
-		content = contentService.findById(contentId);
-		assertEquals(content.getTitle(),title);
-	}
-	
-	@WithMockUser(username="test@google.com", roles= {"USER"})
-	@Test
-	@Transactional
-	@DisplayName("콘텐트 업데이트 POST(정상)")
-	public void updateContentPostTest() throws Exception {
-		
-		Person p = personService.findByEmail("test@google.com");
-		Space space = p.getSpace();
-		Content content = space.getContents().get(0);
-		Long contentId = content.getId();
-		
-		MultiValueMap<String, String> map = new LinkedMultiValueMap<>();
-		map.add("hostEmail", "test@google.com");
-		map.add("title", "changed title"); 
-		map.add("contributors", "changed test contributors");
-		map.add("contents", "changed test contents");
-		
-		mockMvc.perform(post("/space/updatecontent/" + contentId.intValue())
-				.params(map))
-				.andExpect(redirectedUrl("/space"));
-		
-		content = contentService.findById(contentId);
-		
-		assertEquals(content.getTitle(),"changed title");
-		assertEquals(content.getContributors(),"changed test contributors");
-		assertEquals(content.getContents(),"changed test contents");
-	}
-	
-	/*
-	 *  - 대상 메소드:
-	 *  	@RequestMapping(value="/space/deletecontent/{id}", method = RequestMethod.DELETE)
-	 * 		public ModelAndView deleteContent(@PathVariable long id)
-	 */
-	
-	@WithMockUser(username="second@naver.com", roles= {"USER"})
-	@Test
-	@Transactional
-	@DisplayName("콘텐트 삭제 (실패-해당 id의 콘텐트가 현재 접속자의 것이 아닐때)")
-	public void deleteContentWrongUser() throws Exception {
-		Person p = personService.findByEmail("test@google.com");
-		Space space = p.getSpace();
-		Content content = space.getContents().get(0);
-		Long contentId = content.getId();
-		
-		mockMvc.perform(delete("/space/deletecontent/" + contentId.intValue()))
-				.andExpect(status().isBadRequest());
-		
-		assertEquals(contentService.existsById(contentId),true);
-	}
-	
-	@WithMockUser(username="test@google.com", roles= {"USER"})
-	@Test
-	@DisplayName("콘텐트 삭제 (실패-입력된 id에 해당하는 content가 DB에 없을때)")
-	public void deleteContentNonExist() throws Exception{
-		mockMvc.perform(delete("/space/deletecontent/100"))
-				.andExpect(status().isNoContent());
-	}
-	
-	@WithMockUser(username="test@google.com", roles= {"USER"})
-	@Test
-	@Transactional 
-	@DisplayName("콘텐트 삭제 (정상)")
-	public void deleteContentTest() throws Exception {
-		long num = contentService.countAll();
-		
-		Person p = personService.findByEmail("test@google.com");
-		Space space = p.getSpace();
-		Content content = space.getContents().get(0);
-		Long contentId = content.getId();
-		
-		int contentsListSize = space.getContents().size();
-		mockMvc.perform(delete("/space/deletecontent/" + contentId.intValue()))
-				.andExpect(redirectedUrl("/space"));
-		
-		assertEquals(contentsListSize-1, space.getContents().size());
-		assertEquals(num-1,contentService.countAll());
 	}
 }
