+++ src/test/java/com/gunyoung/info/controller/rest/RestfulControllerTest.java
@@ -1,4 +1,4 @@
-package com.gunyoung.info.controller;
+package com.gunyoung.info.controller.rest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
